+++ sandbox/src/main.c
@@ -642,7 +642,7 @@ static const char* asset_dir = "../../../sandbox/assets";
 void SandboxInit() {
 
     LinceApp* app = LinceGetApp();
-    LincePushAssetDir(&app->asset_manager, asset_dir);
+    LinceAppPushAssetDir(asset_dir);
     LinceInitCamera(&DATA.camera, LinceAppGetAspectRatio());
     DATA.camera.zoom = 3.0f;
     DATA.camera_speed = 0.003f;