+++ sandbox/src/scenes/house_scene.c
@@ -39,7 +39,7 @@ void HouseSceneInit(LinceScene* scene){
     scene->data = house_scene;
     
     // Town map
-    char* map_path = LinceFetchAssetPath(&LinceGetApp()->asset_manager, "textures/inside.png");
+    char* map_path = LinceAppFetchAssetPath("textures/inside.png");
     
     house_scene->map =  (LinceTilemap){
         .texture = LinceLoadTexture(